+++ src/app/[locale]/layout.tsx
@@ -197,6 +197,24 @@ export default async function LocaleLayout({
                     defer
                 />
 
+                {/* Google tag (gtag.js) */}
+                <Script
+                    src="https://www.googletagmanager.com/gtag/js?id=G-47QZWWTPZ1"
+                    strategy="afterInteractive"
+                />
+                <Script
+                    id="google-analytics"
+                    strategy="afterInteractive"
+                    dangerouslySetInnerHTML={{
+                        __html: `
+                            window.dataLayer = window.dataLayer || [];
+                            function gtag(){dataLayer.push(arguments);}
+                            gtag('js', new Date());
+
+                            gtag('config', 'G-47QZWWTPZ1');
+                        `,
+                    }}
+                />
             </body>
         </html>
     );